backend/requirements.txt
fastapi
uvicorn
python-multipart
pytesseract
opencv-python
pdf2image
pillow
openai
python-dotenv